Add NGram feature
Turns out "Where len(g) and freq(g) are the number of words in the ngram g and its observed frequency in the historical query logs" is almost the only indicator that ngrams need to consist of words (and not characters!)

Note: IndexSearcher needs to be set to a non-null value in order to enable this feature. `LambdaMARTCompletionTest`, therefore currently does NOT use the NGram feature.